refactor: move conditions into machine pool package
