+++ controllers/machinepool/conditions.go
@@ -1,4 +1,4 @@
-package conditions
+package machinepool
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
